Set tags correctly for release and worker-image workflows
We switched from manually executing `docker buildx bake` to using the
docker metadata and build-push action.
These don't support the format we were using and only the first tag was
published.

Additionally, the metadata action is actually able to do with the
tagging what we were doing manually, so just replaced all that with what
the metadata action supports.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>